[runtime] Store implmap_idx as a guint32 to avoid overflows if the implmap table has more than 64k rows. Fixes mono/mono#59881.
Commit migrated from https://github.com/mono/mono/commit/9369ed6e23ee9e822a3aa2318ed55f083241fbcb

+++ src/mono/mono/metadata/class-internals.h
@@ -111,7 +111,7 @@ struct _MonoMethodPInvoke {
 	gpointer addr;
 	/* add marshal info */
 	guint16 piflags;  /* pinvoke flags */
-	guint16 implmap_idx;  /* index into IMPLMAP */
+	guint32 implmap_idx;  /* index into IMPLMAP */
 };
 
 /*
